clean up aggergate node API
